Added checks while trying to delete directories
Former-commit-id: 4dcebd572439ed45dc49ef577220263741fc6dbd

+++ dagent/script/run_dagent.sh
@@ -2,11 +2,16 @@
 
 SCRIPT_PATH=$(readlink -f $(dirname $0))
 PARENT_SCRIPT_PATH="$(dirname "$SCRIPT_PATH")"
-mkdir -p /var/log/m9k/
-sudo systemctl stop dagent
+sudo mkdir -p /var/log/m9k/
+
+if [ -e "/etc/systemd/system/dagent.service" ]; then
+  sudo systemctl stop dagent
+fi
 
 #create a soft link of the directory, to get the absolute path for starting the service
-sudo rm /usr/local/dagent
+if [ -e "/usr/local/dagent" ]; then
+  sudo rm /usr/local/dagent
+fi
 sudo ln -s $PARENT_SCRIPT_PATH /usr/local
 
 #move the service file  to the /etc/systemd/system/
